% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_fitters.R
\name{add_confidence}
\alias{add_confidence}
\title{Add confidence values to the config}
\usage{
add_confidence(config, values = NULL, func = NULL, normalize = TRUE, ...)
}
\arguments{
\item{config}{phases_config class object}

\item{values}{numeric vector; see \code{fit_with_steps} for details}

\item{func}{function; see \code{fit_with_function} for details}

\item{normalize}{logical; \code{TRUE} is the default; should the confidence values be normalized
so that the maximum value is 1 and other elements are rescaled proportionally; see \code{normalize_values}
for details}

\item{...}{arguments passed to fitter functions}
}
\value{
tibble
}
\description{
Add confidence values to the config
}
\examples{
set.seed(123)
df <- create_toy_dataset()
visits <- df$visits
auto_config <- auto_create_phases_config(visits, phases)
auto_config \%>\% add_confidence(values = .5)
auto_config \%>\% add_confidence(values = .5, normalize = FALSE)
auto_config \%>\% add_confidence(values = c(0.5, 0.5, 0.7, 0.7, rep(1, 6)))

# fit with gaussian distribution:
gauss <- auto_config \%>\% add_confidence(func = dnorm)
gauss <- auto_config \%>\% add_confidence(func = dnorm, mean = -1, sd = 3)

library(ggplot2)
ggplot(
  data = gauss,
  aes(x = time_point, y = confidence)
) +
  geom_point() +
  geom_line() +
  facet_wrap(facets = "phase")
}
\seealso{
\code{\link{fit_with_steps}}, \code{\link{fit_with_function}}, \code{\link{normalize_values}}
}
