% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_truth_labels_and_config.R
\name{auto_create_phases_config}
\alias{auto_create_phases_config}
\title{Automatically creates config object based on input data}
\usage{
auto_create_phases_config(df, phases_column)
}
\arguments{
\item{df}{tibble}

\item{phases_column}{tidy-select specification of the column with phases}
}
\description{
Automatically creates config object based on input data
}
