% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{save_bipolar_config}
\alias{save_bipolar_config}
\alias{set_inputs_location}
\alias{set_outputs_location}
\title{Saves bipolar configuration object to YAML file}
\usage{
save_bipolar_config(config, path)

set_inputs_location(config, path)

set_outputs_location(config, path)
}
\arguments{
\item{config}{bipolar.config-class object}

\item{path}{character; where to store the file}
}
\description{
\code{save_bipolar_config} saves bipolar configuration object in a YAML format.
\code{set_inputs_location} and \code{set_outputs_location} are convenient methods to modify the
paths of a configuration object
}
\seealso{
Other bipolar-config: 
\code{\link{create_bipolar_config}()},
\code{\link{read_bipolar_config}()},
\code{\link{save_output}()}
}
\concept{bipolar-config}
