% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_f1_metrics}
\alias{calculate_f1_metrics}
\title{Calculate F1 score for binary and multiclass problem}
\usage{
calculate_f1_metrics(
  true_y,
  predicted_y,
  mode = "multiclass",
  average = "macro"
)
}
\arguments{
\item{true_y}{factor containing true labels}

\item{predicted_y}{factor containing predicted labels}

\item{mode}{character specifying binary or multiclass label problem.}

\item{average}{specifies how to summarize overall F1.
* `macro` will take unweighted mean of F1 scores for each class,
* any other value will exclude overall summary from the returned data frame.}
}
\value{
Dataframe with F1 results.
}
\description{
Calculate F1 score for binary and multiclass problem
}
\examples{
\dontrun{
phases <- c("euthymia", "depression", "mania")
yt <- factor(sample(x = phases, replace = TRUE, size = 100), levels = phases)
yp <- factor(sample(x = phases, replace = TRUE, size = 100), levels = phases)
calculate_f1_metrics(yt, yp)
calculate_f1_metrics(yt, yp, average = FALSE)
}
}
