% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_fitters.R
\name{conf_fitters}
\alias{conf_fitters}
\alias{fit_with_steps}
\alias{fit_with_function}
\title{Fitting time window with steps}
\usage{
fit_with_steps(
  values,
  .time_window = time_window(c(-DEFAULT_PHASE_DAYS_BEFORE, DEFAULT_PHASE_DAYS_AFTER)),
  ...
)

fit_with_function(
  f,
  .time_window = time_window(c(-DEFAULT_PHASE_DAYS_BEFORE, DEFAULT_PHASE_DAYS_AFTER)),
  ...
)
}
\arguments{
\item{values}{numeric vector; must be length 1 or the same as \code{time_window}}

\item{.time_window}{two-element numeric vector; the first element is the beginning of a time window,
the other element is the end of it.}

\item{...}{arguments passed to the function \code{f}}

\item{f}{function; a function returning a numeric vector of the same lenght as \code{time_window}}
}
\value{
tibble; the first column is a sequence of days according to the \code{time_window} definition,
the second column is populated with \code{values}
}
\description{
Fitting time window with steps
}
\details{
For \code{fit_with_steps} - in case \code{values} is one element vector it will be
repeated to fit the length of \code{time_window}
}
\examples{
fit_with_steps(0.5)
fit_with_steps(c(0.5, 0.5, 0.7, 0.7, 1), time_window(c(-2, 2)))
# fit time window with values from a gaussian density function
fit_with_function(dnorm, mean = -1, sd = 3)
}
