% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{create_bipolar_config}
\alias{create_bipolar_config}
\title{Creates bipolar configuration object.}
\usage{
create_bipolar_config(inputs = NULL, outputs = NULL)
}
\arguments{
\item{inputs}{character; path to the location on disk where inputs are stored}

\item{outputs}{character; path to the location on disk where experiment outputs will be stored}
}
\value{
bipolar.config-class object
}
\description{
Bipolar configuration requires defining paths to two directories:
\item{inputs} - the directory where input datasets are stored
\item{outputs} - where all experimentation/modeling results will be stored
}
\details{
By default, \code{inputs} and \code{outputs} arguments are \code{NULL}. In this case inputs
and outputs directories will be created as subdirectories of the current working directory
(\code{getwd()} is used.)
A side effect of creation of a config object is setting two environment variables, namely
\code{INPUTS_DIR} and \code{OUTPUTS_DIR}. The main advantage of this effect is that inputs/outputs
locations can be obtained easily.
}
\seealso{
Other bipolar-config: 
\code{\link{read_bipolar_config}()},
\code{\link{save_bipolar_config}()},
\code{\link{save_output}()}
}
\concept{bipolar-config}
