% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_truth_labels_and_config.R
\name{create_time_window_config}
\alias{create_time_window_config}
\title{Automatically creates config object based on input data}
\usage{
create_time_window_config(
  df,
  phases_column,
  extrapolation_interval = get_default_intervals_for_phases()
)
}
\arguments{
\item{extrapolation_interval}{a numeric vector of length 2 specifying the
time window interval. The first element should be > 0 and specifies
how many days before the visit day shall the time window span.
The second element should be > 0 and specifies how many days after the visit
shall the time window span.}
}
\description{
Automatically creates config object based on input data
}
