% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{divide_train_test}
\alias{divide_train_test}
\title{Creates explicitly cross-validation train/test splits indices for observations.}
\usage{
divide_train_test(i, folds)
}
\arguments{
\item{i}{index of a fold considered when invoking this function.}

\item{folds}{list of length K. Each ith element of the list contains a numeric
vector denoting indices of observations assigned to the ith fold.}
}
\value{
List with two named elements `train` and `test`, each containing
vector with indices of observations that are categorized as *training data*
or *test data*.
}
\description{
Creates explicitly cross-validation train/test splits indices for observations.
}
\examples{
\dontrun{
folds_info <- list(`0`=c(1, 6, 7, 19, 20), `1`=c(4, 5, 8, 9, 11),
                   `2`=c(2, 3, 12, 13, 14), `3`=c(10, 15, 16, 17, 18))
divide_train_test(1, folds_info)
}
}
