% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_truth_labels_and_config.R
\name{expand_ground_truth_period}
\alias{expand_ground_truth_period}
\title{Extrapolates ground truth labels according to phases configuration}
\usage{
expand_ground_truth_period(
  d,
  config,
  phases_col,
  visit_date_col,
  patient_id_col
)
}
\arguments{
\item{d}{data.frame (tibble) with original visit dates}

\item{config}{period configuration class object;}

\item{phases_col}{which column in \code{visits} stores phases; tidy-selection is available}

\item{visit_date_col}{which column in \code{visits} stores visit date; tidy-selection is available}

\item{patient_id_col}{which column in \code{visits} stores the ID of a patient; tidy-selection is available}
}
\value{
data.frame (tibble) with extra rows covering the interval between
before and after dates from \code{config}
}
\description{
Phases with their corresponding before and after limits should be provided as
\code{config} argument. Phases (labels) will be extrapolated to cover the whole period. Technically,
rows will be multiplicated and the original label will be extrapolated on the new rows.

New rows will be added for every combination of phase and visit date.
}
\examples{
set.seed(123)
df <- create_toy_dataset()
visits <- df$visits
auto_config <- auto_create_phases_config(visits, phases)

#insert new rows for each record
expand_ground_truth_period(df, auto_config, phases, visit_date)
}
