% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_structure.R
\name{get_acoustic_params}
\alias{get_acoustic_params}
\title{A function to retrieve colnames of features that are identified with
prefix "X_".}
\usage{
get_acoustic_params(names, original_names = TRUE)
}
\arguments{
\item{names}{character vector containing raw column names}

\item{original_names}{boolean, whether to keep raw names (if FALSE) or
remove "X_" prefix (if TRUE)}
}
\value{
character vector
}
\description{
A function to retrieve colnames of features that are identified with
prefix "X_".
}
\examples{
\dontrun{
df <- data.frame("a" = 1, "b" = 2, "X_ffa" = -99, "X_long_name" = NA)
get_acoustic_params(colnames(df))
}

}
