% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_truth_labels_and_config.R
\name{get_default_intervals_for_phases}
\alias{get_default_intervals_for_phases}
\title{Gets the default time window borders}
\usage{
get_default_intervals_for_phases()
}
\description{
Gets the default time window borders
}
