% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_truth_labels_and_config.R
\name{get_phase_interval_from_config}
\alias{get_phase_interval_from_config}
\title{Gets time window borders for selected phase}
\usage{
get_phase_interval_from_config(config, .phase)
}
\arguments{
\item{config}{time window configuration (\code{phases_config} class object)}

\item{.phase}{character; selected phase}
}
\description{
Gets time window borders for selected phase
}
