% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_parameters.R
\name{get_sample_call_parameters}
\alias{get_sample_call_parameters}
\title{Gets sample parameters of calls}
\usage{
get_sample_call_parameters()
}
\value{
character vector of selected parameters
}
\description{
Gets sample parameters of calls
}
