% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sample_ds}
\alias{get_sample_ds}
\alias{get_sample_csv}
\alias{get_sample_visits}
\alias{get_sample_mobile_recordings}
\alias{get_sample_mobile_chunks}
\title{Read sample dataset}
\usage{
get_sample_csv(dataset_name)

get_sample_visits()

get_sample_mobile_recordings()

get_sample_mobile_chunks()
}
\arguments{
\item{dataset_name}{character;}
}
\value{
tibble
}
\description{
Read sample dataset
}
\concept{get_sample_data}
