% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_truth_labels_and_config.R
\name{get_unique_phases}
\alias{get_unique_phases}
\title{Extracts uniqe phases from a dataframe}
\usage{
get_unique_phases(df, phases_column)
}
\arguments{
\item{df}{tibble}

\item{phases_column}{tidy-select specification of the column with phases}
}
\description{
Extracts uniqe phases from a dataframe
}
