% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{new_bipolar_config}
\alias{new_bipolar_config}
\title{User-friendly bipolar config object creator}
\usage{
new_bipolar_config(x)
}
\arguments{
\item{x}{should be a list with two named elements: inputs and outputs}
}
\value{
bipolar.config-class object
}
\description{
User-friendly bipolar config object creator
}
