% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_fitters.R
\name{normalize_values}
\alias{normalize_values}
\title{Normalize values}
\usage{
normalize_values(x)
}
\arguments{
\item{x}{numeric; all its elements must be non-negative}
}
\value{
vector of the same length as \code{x}
}
\description{
The function re-scales values of non-negative numeric vector so the maximum is 1
}
\details{
Notice, that we cannot use min-max rescaling because this method returns a vector in 0-1
range. This means that the minimal element of \code{x} would be rescaled to 0.
}
