% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{save_output}
\alias{save_output}
\alias{load_results}
\alias{list_results}
\title{Read/save model outputs}
\usage{
save_output(x, name)

load_results(name)

list_results()
}
\arguments{
\item{x}{R object to be saved}

\item{name}{character; for saving - the name of the resulting file; for reading - the name of the
resource to be loaded}
}
\description{
When experimenting a lot of different parameter sets may be tested as well as input data. Storing
the results for later analysis and deeper insights is a key aspect of modeling process and
experimentation in general. Here, the two functions are responsible for saving and reading the
outcomes from experimentation.
}
\details{
Results are stored as RDS files via base R \code{saveRDS} function. For reading saved objects
back the reverse \code{readRDS} function is used.
If the outputs directory doesn't exist it will be created.
\code{list_results} function returns a data frame with information about all results stored in the
outputs directory. You can use it to get proper names for the assets You want to load.
}
\seealso{
Other bipolar-config: 
\code{\link{create_bipolar_config}()},
\code{\link{read_bipolar_config}()},
\code{\link{save_bipolar_config}()}
}
\concept{bipolar-config}
