% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{read_bipolar_config}
\alias{read_bipolar_config}
\title{Reads bipolar configuration from a YAML file}
\usage{
read_bipolar_config(path = NULL)
}
\arguments{
\item{path}{character; path to the yaml file}
}
\value{
bipolar.config-class object
}
\description{
Reads bipolar configuration from a YAML file
}
\seealso{
Other bipolar-config: 
\code{\link{create_bipolar_config}()},
\code{\link{save_bipolar_config}()},
\code{\link{save_output}()}
}
\concept{bipolar-config}
