% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Runs shiny dashboard}
\usage{
run_shiny(.port = 7777)
}
\arguments{
\item{.port}{integer; which port to allocate? 7777 is the default}
}
\description{
Runs shiny dashboard
}
