% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{splitter_visits}
\alias{splitter_visits}
\title{The function to create train/test split in scenario-visit-out (SVO)
scenario.}
\usage{
splitter_visits(.data, visit_out = c("last", "random"), min_visits = 1)
}
\arguments{
\item{.data}{final dataframe for modeling purposes}

\item{visit_out}{a character indicator, either "random" or "last"}

\item{min_visits}{numeric, how many visits at least
patients must have had to be considered for inclusion in the test set}
}
\value{
a list of two lists named `train` and `test`, each list
including integers being indices of observations chosen to be in the
respective train/test split
}
\description{
The function to create train/test split in scenario-visit-out (SVO)
scenario.
}
\examples{
\dontrun{
names <- c("patient_id", "visit_date", "label", "variable")

df_example <- rbind(
  setNames(data.frame(1, NA, NA, rnorm(20)), names),
  setNames(data.frame(2, "2021-03-17", "depression", rnorm(7)), names),
  setNames(data.frame(2, "2021-04-18", "depression", rnorm(6)), names),
  setNames(data.frame(2, NA, NA, rnorm(24)), names),
  setNames(data.frame(3, "2022-10-26", "euthymia", rnorm(11)), names),
  setNames(data.frame(3, "2022-11-27", "mixed", rnorm(9)), names),
  setNames(data.frame(3, "2022-12-28", "mixed", rnorm(5)), names),
  setNames(data.frame(3, NA, NA, rnorm(30)), names)
)

splits_random <- bipolar::splitter_visits(df_example, "random")
df_example[splits_random[[1]]$train, ]
df_example[splits_random[[1]]$test, ]
}
}
