% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_time_window.R
\name{time_window}
\alias{time_window}
\title{Simple class for time window sequencer}
\usage{
time_window(x)
}
\arguments{
\item{x}{two element integer vector specifying limits of a window, for example \code{c(-7, 2)}}
}
\value{
sequence of integer values from \code{x[1]} to \code{x[2]}
}
\description{
Simple class for time window sequencer
}
\examples{
time_window(c(-2, 2)) # returns sequence -2 -1 0 1 2
}
