% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{transform_label_cgi}
\alias{transform_label_cgi}
\title{Transform column into cgi label}
\usage{
transform_label_cgi(dataset, label_name)
}
\arguments{
\item{dataset}{data.frame}

\item{label_name}{character}
}
\value{
dataset
}
\description{
Transform column into cgi label
}
\examples{
visits <- get_sample_visits()
transform_label_healthy_unhealthy(visits, "hamd_ymrs")
}
