% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{transform_label_custom}
\alias{transform_label_custom}
\title{Transforms column with states into custom label}
\usage{
transform_label_custom(dataset, label_name)
}
\arguments{
\item{dataset}{data.frame}

\item{label_name}{character}
}
\value{
data.frame
}
\description{
Transforms column with states into custom label
}
