% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{transform_label_healthy_unhealthy}
\alias{transform_label_healthy_unhealthy}
\title{Transforms column into label with healthy-unhealthy values}
\usage{
transform_label_healthy_unhealthy(dataset, label_name)
}
\arguments{
\item{dataset}{data.frame}

\item{label_name}{character; the name of a column storing state values (mania, depression etc.)}
}
\value{
data.frame
}
\description{
Transforms column into label with healthy-unhealthy values
}
\examples{
visits <- get_sample_visits()
transform_label_healthy_unhealthy(visits, "hamd_ymrs")
}
