% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{transform_label_symptoms}
\alias{transform_label_symptoms}
\title{Transform column into multilabels hamilton and young}
\usage{
transform_label_symptoms(dataset)
}
\arguments{
\item{dataset}{data.frame}
}
\value{
dataset
}
\description{
Transform column into multilabels hamilton and young
}
\examples{
visits <- get_sample_visits()
transform_label_symptoms(visits )
}
