% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{weighted_FS}
\alias{weighted_FS}
\title{Calculate top n features using weighted feature selection}
\usage{
weighted_FS(dataset, number_of_features, target_type, start_column, end_column)
}
\arguments{
\item{dataset}{data.frame (dataframe contains columns with acoustic features, assigned labels - as 'label' and  confidence path as 'confidence')}

\item{number_of_features}{integer}

\item{target_type}{character c('mania', 'depression', 'all')}

\item{start_column}{character (starting column from range of required columns, ex: 'pcm_LOGenergy_sma' )}

\item{end_column}{character (ending column from range of required columns, ex: 'pcm_LOGenergy_sma' 'pcm_fftMag_mfcc_12_')}
}
\value{
list
}
\description{
Calculate top n features using weighted feature selection
}
\examples{
\dontrun{
weighted_FS(dataset = model_df , number_of_features = 10, target_type = 'mania')
}
}
